+++ packages/chakra-ui/src/BaseInputTemplate/BaseInputTemplate.tsx
@@ -10,6 +10,7 @@ import {
   RJSFSchema,
   StrictRJSFSchema,
 } from '@rjsf/utils';
+import { SchemaExamples } from '@rjsf/core';
 
 import { Field } from '../components/ui/field';
 import { getChakra } from '../utils';
@@ -85,15 +86,7 @@ export default function BaseInputTemplate<
       {options.allowClearTextInputs && !readonly && !disabled && value && (
         <ClearButton registry={registry} onClick={onClear} />
       )}
-      {Array.isArray(schema.examples) ? (
-        <datalist id={examplesId(id)}>
-          {(schema.examples as string[])
-            .concat(schema.default && !schema.examples.includes(schema.default) ? ([schema.default] as string[]) : [])
-            .map((example: any) => {
-              return <option key={example} value={example} />;
-            })}
-        </datalist>
-      ) : null}
+      <SchemaExamples id={id} schema={schema} />
     </Field>
   );
 }