+++ packages/fluentui-rc/src/BaseInputTemplate/BaseInputTemplate.tsx
@@ -10,6 +10,7 @@ import {
   StrictRJSFSchema,
   labelValue,
 } from '@rjsf/utils';
+import { SchemaExamples } from '@rjsf/core';
 
 const useStyles = makeStyles({
   input: {
@@ -99,15 +100,7 @@ export default function BaseInputTemplate<
       {options.allowClearTextInputs && !readonly && !disabled && value && (
         <ClearButton registry={registry} onClick={_onClear} />
       )}
-      {Array.isArray(schema.examples) && (
-        <datalist id={examplesId(id)}>
-          {(schema.examples as string[])
-            .concat(schema.default && !schema.examples.includes(schema.default) ? ([schema.default] as string[]) : [])
-            .map((example: any) => {
-              return <option key={example} value={example} />;
-            })}
-        </datalist>
-      )}
+      <SchemaExamples id={id} schema={schema} />
     </>
   );
 }