+++ packages/react-bootstrap/src/BaseInputTemplate/BaseInputTemplate.tsx
@@ -9,6 +9,7 @@ import {
   RJSFSchema,
   StrictRJSFSchema,
 } from '@rjsf/utils';
+import { SchemaExamples } from '@rjsf/core';
 
 export default function BaseInputTemplate<
   T = any,
@@ -77,15 +78,7 @@ export default function BaseInputTemplate<
         <ClearButton registry={registry} onClick={_onClear} />
       )}
       {children}
-      {Array.isArray(schema.examples) ? (
-        <datalist id={examplesId(id)}>
-          {(schema.examples as string[])
-            .concat(schema.default && !schema.examples.includes(schema.default) ? ([schema.default] as string[]) : [])
-            .map((example: any) => {
-              return <option key={example} value={example} />;
-            })}
-        </datalist>
-      ) : null}
+      <SchemaExamples id={id} schema={schema} />
     </>
   );
 }