+++ packages/shadcn/src/BaseInputTemplate/BaseInputTemplate.tsx
@@ -8,6 +8,7 @@ import {
   StrictRJSFSchema,
 } from '@rjsf/utils';
 import { ChangeEvent, FocusEvent, MouseEvent, useCallback } from 'react';
+import { SchemaExamples } from '@rjsf/core';
 
 import { Input } from '../components/ui/input';
 import { cn } from '../lib/utils';
@@ -86,15 +87,7 @@ export default function BaseInputTemplate<
         <ClearButton onClick={_onClear} registry={registry} />
       )}
       {children}
-      {Array.isArray(schema.examples) ? (
-        <datalist id={examplesId(id)}>
-          {(schema.examples as string[])
-            .concat(schema.default && !schema.examples.includes(schema.default) ? ([schema.default] as string[]) : [])
-            .map((example: any) => {
-              return <option key={example} value={example} />;
-            })}
-        </datalist>
-      ) : null}
+      <SchemaExamples id={id} schema={schema} />
     </div>
   );
 }